[flue] Extract style-guide workspace hydration into lib/style-guide-hydration.ts
