+++ .flue/workflows/style-guide-review.ts
@@ -16,9 +16,11 @@ import {
 	getDefaultWorkspace,
 	getShellSandbox,
 } from "../connectors/cloudflare-shell";
+import { hydrateStyleGuideWorkspace } from "../lib/style-guide-hydration";
 import {
 	assignFindingIds,
 	StyleGuideResultFromModelSchema,
+	type StyleGuideResult,
 } from "../lib/style-guide-results";
 export type {
 	StyleGuideFinding,
@@ -27,31 +29,12 @@ export type {
 
 export const route: WorkflowRouteHandler = async (_c, next) => next();
 
-// Only review docs/partials/changelog MDX
-const REVIEWABLE_PATH_RE = /^src\/content\/(docs|partials|changelog)\/.+\.mdx$/;
-
 interface StyleGuideReviewPayload {
 	number: number;
 	diffDir: string;
 	commentsPath: string;
 	/** When set, review only this file. Used by orchestrator fan-out. */
 	filename?: string;
-}
-
-interface ManifestEntry {
-	filename: string;
-	status: string;
-	additions: number;
-	deletions: number;
-	changes: number;
-	patch_key: string | null;
-}
-
-interface PullRequestMetadata {
-	number: number;
-	title: string;
-	base: string;
-	head: string;
 }
 
 export async function run({ id: runId, init, payload, env }: FlueContext) {
@@ -115,11 +98,30 @@ async function runImpl({
 		action: "started",
 	});
 
-	// ── 1. Fast-fail if no diff in R2 ─────────────────────────────────────────
-	const manifestObj = await bucket.get(`${input.diffDir}/manifest.json`);
-	if (!manifestObj) {
+	// ── 1. Hydrate workspace from R2 ──────────────────────────────────────────
+	console.log({
+		message: `Style-guide review hydrating workspace: PR #${input.number}${input.filename ? ` — ${input.filename}` : ""}`,
+		event: "style_guide_review",
+		number: input.number,
+		filename: input.filename ?? null,
+		diffDir: input.diffDir,
+		runId,
+		action: "hydration_start",
+	});
+
+	const hydration = await hydrateStyleGuideWorkspace(bucket, workspace, {
+		diffDir: input.diffDir,
+		commentsPath: input.commentsPath,
+		filename: input.filename,
+		prNumber: input.number,
+	});
+
+	if (!hydration) {
+		const reason = input.filename
+			? "No reviewable documentation files changed."
+			: "No diff files found in R2.";
 		console.log({
-			message: `Style-guide review: no diff files found in R2 for PR #${input.number}`,
+			message: `Style-guide review: ${reason} PR #${input.number}${input.filename ? ` — ${input.filename}` : ""}`,
 			event: "style_guide_review",
 			number: input.number,
 			diffDir: input.diffDir,
@@ -129,145 +131,13 @@ async function runImpl({
 		});
 		return {
 			findings: [],
-			summary: "No diff files found in R2.",
-			reviewedFiles: [],
-		} satisfies StyleGuideResult;
-	}
-	const manifestText = await manifestObj.text();
-	const manifest = JSON.parse(manifestText) as ManifestEntry[];
-	const reviewedFiles = input.filename
-		? manifest.some(
-				(f) =>
-					f.filename === input.filename && REVIEWABLE_PATH_RE.test(f.filename),
-			)
-			? [input.filename]
-			: []
-		: manifest
-				.filter((f) => REVIEWABLE_PATH_RE.test(f.filename))
-				.map((f) => f.filename);
-	if (reviewedFiles.length === 0) {
-		return {
-			findings: [],
-			summary: "No reviewable documentation files changed.",
+			summary: reason,
 			reviewedFiles: [],
 		} satisfies StyleGuideResult;
 	}
 
-	// ── 2. Read PR metadata from R2 ────────────────────────────────────────────
-	const prObj = await bucket.get(`${input.diffDir}/pr.json`);
-	const prText = prObj ? await prObj.text() : null;
-	const pullRequest = prText
-		? (JSON.parse(prText) as PullRequestMetadata)
-		: {
-				number: input.number,
-				title: "",
-				base: "",
-				head: "",
-			};
-
-	// ── 3. Populate workspace from R2 before init ─────────────────────────────
-	// When a specific filename is requested, load only what that child run
-	// needs: the manifest, pr.json, comments, and the single patch file.
-	// This avoids duplicating every sibling patch into each concurrent DO.
-	// When no filename is given (full-PR mode), fall back to bulk hydration.
-
-	const [referenceObjects, skillObj] = await Promise.all([
-		bucket.list({ prefix: ".agents/reference/style-guide/" }),
-		bucket.get(".agents/skills/style-guide-review/SKILL.md"),
-	]);
-	if (!skillObj) {
-		throw new Error(
-			"Missing .agents/skills/style-guide-review/SKILL.md in DOCS_FLUE_BUCKET. " +
-				"For local dev, run `pnpm run flue:sync-agents:local` before invoking the workflow.",
-		);
-	}
-
-	// Determine which diff objects to load.
-	// Targeted mode (filename set): manifest.json and pr.json are already in
-	// memory from steps 1 and 2 — reuse their text rather than re-fetching.
-	// Only fetch comments and the single patch_key for the requested file.
-	// Full mode (no filename): load everything under the diffDir prefix.
-	//
-	// diffResults shape: { key, text }[]  (same as the full-mode path below)
-	let diffKeysToLoad: string[];
-	let cachedDiffResults: { key: string; text: string }[] = [];
-	if (input.filename) {
-		const entry = manifest.find((f) => f.filename === input.filename);
-		const patchKey = entry?.patch_key ?? null;
-
-		// Reuse already-read strings — manifestText and prText are captured above.
-		cachedDiffResults = [
-			{ key: `${input.diffDir}/manifest.json`, text: manifestText },
-			...(prText !== null
-				? [{ key: `${input.diffDir}/pr.json`, text: prText }]
-				: []),
-		];
-
-		// Only fetch comments + the patch file fresh.
-		diffKeysToLoad = [input.commentsPath, ...(patchKey ? [patchKey] : [])];
-	} else {
-		const all = await bucket.list({ prefix: `${input.diffDir}/` });
-		diffKeysToLoad = all.objects.map((o) => o.key);
-	}
-
-	// Read all reference files and any remaining diff files in parallel.
-	const [referenceResults, ...fetchedDiffResults] = await Promise.all([
-		Promise.all(
-			referenceObjects.objects.map(async (obj) => ({
-				key: obj.key,
-				text: (await (await bucket.get(obj.key))?.text()) ?? "",
-			})),
-		),
-		...diffKeysToLoad.map(async (key) => ({
-			key,
-			text: (await (await bucket.get(key))?.text()) ?? "",
-		})),
-	]);
-
-	// Merge cached (already-fetched) entries with freshly-fetched ones.
-	const diffResults = [...cachedDiffResults, ...fetchedDiffResults];
-
-	console.log({
-		message: `Style-guide review hydrating workspace: PR #${input.number}${input.filename ? ` — ${input.filename}` : ""}`,
-		event: "style_guide_review",
-		number: input.number,
-		filename: input.filename ?? null,
-		diffDir: input.diffDir,
-		diffObjects: cachedDiffResults.length + diffKeysToLoad.length,
-		referenceObjects: referenceObjects.objects.length,
-		runId,
-		action: "hydration_start",
-	});
-
-	// Pre-create common parent directories before parallel writes. Otherwise
-	// concurrent writeFile calls can race while creating the same directory rows
-	// in the cf-shell workspace SQLite table.
-	for (const dir of [
-		"/.agents/skills/style-guide-review",
-		"/.agents/reference/style-guide/always",
-		"/.agents/reference/style-guide/conditional",
-		"/.agents/reference/style-guide/components",
-		`/${input.diffDir}`,
-	]) {
-		await workspace.mkdir(dir, { recursive: true });
-	}
-
-	// Write everything to workspace in parallel.
-	await Promise.all([
-		// Skill file
-		workspace.writeFile(
-			"/.agents/skills/style-guide-review/SKILL.md",
-			await skillObj.text(),
-		),
-		// All reference files (preserving subdirectory structure)
-		...referenceResults.map((r) =>
-			r.text ? workspace.writeFile(`/${r.key}`, r.text) : Promise.resolve(),
-		),
-		// Selected diff files (manifest, pr.json, comments, and patch)
-		...diffResults.map((r) =>
-			r.text ? workspace.writeFile(`/${r.key}`, r.text) : Promise.resolve(),
-		),
-	]);
+	const { pullRequest, reviewedFiles, diffObjects, referenceObjects } =
+		hydration;
 
 	console.log({
 		message: `Style-guide review workspace ready: PR #${input.number}${input.filename ? ` — ${input.filename}` : ""}`,
@@ -275,13 +145,13 @@ async function runImpl({
 		number: input.number,
 		filename: input.filename ?? null,
 		diffDir: input.diffDir,
-		diffObjects: cachedDiffResults.length + diffKeysToLoad.length,
-		referenceObjects: referenceObjects.objects.length,
+		diffObjects,
+		referenceObjects,
 		runId,
 		action: "hydration_complete",
 	});
 
-	// ── 4. Init harness ───────────────────────────────────────────────────────
+	// ── 2. Init harness ───────────────────────────────────────────────────────
 	const agent = createAgent(() => ({
 		sandbox: getShellSandbox({ workspace, loader }),
 		model: "cloudflare/@cf/moonshotai/kimi-k2.7-code",
@@ -289,13 +159,14 @@ async function runImpl({
 	}));
 	const harness = await init(agent);
 
-	// ── 5. Run the skill ───────────────────────────────────────────────────────
+	// ── 3. Run the skill ───────────────────────────────────────────────────────
 	const session = await harness.session(
 		`style-guide-review:${input.number}:${runId}`,
 	);
 
-	// Use structured result mode so flue injects finish/give_up tools and loops until the
-	// model calls finish — works reliably across models that don't self-terminate.
+	// Use structured result mode so flue injects finish/give_up tools and loops
+	// until the model calls finish — works reliably across models that don't
+	// self-terminate.
 	const skillResult = await session.skill("style-guide-review", {
 		result: StyleGuideResultFromModelSchema,
 		args: {
